requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
